removed announcement detail page, now redundant. Added the code to the placeholder page just in case
